Aegis-Hire-App/backend/auth.py
from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
import os
from dotenv import load_dotenv
from supabase import Client
from database import supabase
from typing import Optional
import jwt
from datetime import datetime, timedelta

# Load environment variables
load_dotenv()

# Security scheme for API endpoints
security = HTTPBearer()

def get_current_user(credentials: HTTPAuthorizationCredentials = Depends(security)) -> Optional[dict]:
    """
    Get the current authenticated user from the JWT token.
    In a real implementation, this would verify the token with Supabase Auth.
    """
    token = credentials.credentials
    
    # In a real implementation, we would verify the token with Supabase Auth
    # For now, we'll implement a simple mock authentication
    try:
        # This is a placeholder - in a real app, you'd verify the token with Supabase
        # decoded_token = supabase.auth.get_user(token)
        # return decoded_token.user
        
        # For now, return a mock user
        return {"id": "mock_user_id", "email": "user@example.com", "role": "admin"}
    except Exception:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid authentication credentials",
            headers={"WWW-Authenticate": "Bearer"},
        )


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    """
    Create an access token for the user.
    In a real implementation, this would be handled by Supabase Auth.
    """
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    
    # In a real implementation, Supabase would handle token creation
    # encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    # return encoded_jwt
    
    # For now, return the data as a mock token
    return {"access_token": "mock_token", "token_type": "bearer", **to_encode}
